Revert "xrga"
This reverts commit 77c40e64a99cba85cc258f413a75bc49389c3738.

+++ src/xrGame/script_game_object3.cpp
@@ -1468,7 +1468,7 @@ u8 CScriptGameObject::GetRestrictionType()
 	if (restr)
 		return restr->m_space_restrictor_type;
 
-	return u8(-1);
+	return (-1);
 }
 
 void CScriptGameObject::SetRestrictionType(u8 typ)
